Restructur files and folders
