+++ main.rb
@@ -1,0 +1,8 @@
+require_relative 'src/app'
+
+def main
+  app = App.new
+  app.run
+end
+
+main
