+++ src/classes/label.rb
