+++ src/classes/music_album.rb
@@ -1,4 +1,4 @@
-require_relative '../items'
+require_relative 'items'
 require 'json'
 
 class MusicAlbum < Items
