--- src/main.rb
@@ -1,8 +1,0 @@
-require_relative 'app'
-
-def main
-  app = App.new
-  app.run
-end
-
-main
